remove unused options/fixtures + separate ds/storage fixtures
