+++ hub/constants.py
@@ -49,3 +49,13 @@ UUID_SHIFT_AMOUNT = 128 - (8 * ENCODING_DTYPE(1).itemsize)
 PYTEST_MEMORY_PROVIDER_BASE_ROOT = "mem://hub_pytest"
 PYTEST_LOCAL_PROVIDER_BASE_ROOT = "/tmp/hub_pytest/"  # TODO: may fail for windows
 PYTEST_S3_PROVIDER_BASE_ROOT = "s3://hub-2.0-tests/"
+
+
+# pytest options
+MEMORY_OPT = "--memory-skip"
+LOCAL_OPT = "--local"
+S3_OPT = "--s3"
+S3_PATH_OPT = "--s3-path"
+CACHE_OPT = "--cache-chains"
+CACHE_ONLY_OPT = "--cache-chains-only"
+KEEP_STORAGE_OPT = "--keep-storage"
