+++ setup.cfg
@@ -1,7 +1,5 @@
 [tool:pytest]
 addopts = -s
-markers =
-    full_benchmark: mark a benchmark that will take a long time as "full". in order to run this benchmark, the `--full-benchmarks` option is required.
 
 [mypy]
 warn_redundant_casts = True
